[API] Sessions: create / destroy session + access

+++ server/routes/Courses/courses.js
@@ -4,7 +4,7 @@ const Course = require("../../models/Courses/Course")
 const CourseShared = require("../../models/Courses/CourseShared")
 const User = require("../../models/Users/User")
 const Role = require("../../models/Users/Role")
-const auth = require("../../middleware/auth");
+const auth = require("../../middleware/auth")
 
 /**
  * @method - POST
@@ -13,24 +13,28 @@ const auth = require("../../middleware/auth");
  */
 router.post("/", auth, async (req, res) => {
     try{
-        const newCourse = new Course(req.body)
-        newCourse.save()
-        
-        User.findOne({_id: mongoose.Types.ObjectId(req.body.owner_id)}, function(err, user) {
-            if (err) return console.log(err)
-            user.courses.push(newCourse._id)
-            user.save()
-        });
+        if(req.session.isAuth){
+            const newCourse = new Course(req.body)
+            newCourse.save()
+            
+            User.findOne({_id: mongoose.Types.ObjectId(req.body.owner_id)}, function(err, user) {
+                if (err) return console.log(err)
+                user.courses.push(newCourse._id)
+                user.save()
+            });
 
-        const role_admin = await Role.findById("618702283f5059816c261d99")
-        const newCourseShared = new CourseShared({
-            course_id: newCourse._id,
-            user_id: newCourse.owner_id,
-            roles: [role_admin._id]
-        })
-        newCourseShared.save()
-        
-        res.redirect("/api/courses");
+            const role_admin = await Role.findById("618702283f5059816c261d99")
+            const newCourseShared = new CourseShared({
+                course_id: newCourse._id,
+                user_id: newCourse.owner_id,
+                roles: [role_admin._id]
+            })
+            newCourseShared.save()
+            
+            res.redirect("/api/courses")
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     }catch(err){
         res.status(500).json(err)
     }
@@ -43,8 +47,12 @@ router.post("/", auth, async (req, res) => {
  */
 router.put("/:id", auth, async (req, res) => {
     try{
-        const updatedCourse = await Course.findByIdAndUpdate(req.params.id, {$set: req.body}, {new: true})
-        res.status(200).json(updatedCourse)
+        if(req.session.isAuth){
+            const updatedCourse = await Course.findByIdAndUpdate(req.params.id, {$set: req.body}, {new: true})
+            res.status(200).json(updatedCourse)
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     }catch(err){
         res.status(500).json(err)
     }
@@ -57,16 +65,20 @@ router.put("/:id", auth, async (req, res) => {
  */
 router.delete("/:id/:user", auth, async (req, res) => {
     try{
-        await Course.findByIdAndDelete(req.params.id)
+        if(req.session.isAuth){
+            await Course.findByIdAndDelete(req.params.id)
 
-        await User.findByIdAndUpdate(req.params.user, 
-            { $pull: { courses: mongoose.Types.ObjectId(req.params.id)} },
-            { new: true }
-        )
+            await User.findByIdAndUpdate(req.params.user, 
+                { $pull: { courses: mongoose.Types.ObjectId(req.params.id)} },
+                { new: true }
+            )
 
-        await CourseShared.deleteOne({ course_id: req.params.id });
+            await CourseShared.deleteOne({ course_id: req.params.id });
 
-        res.redirect("/api/courses");
+            res.redirect("/api/courses")
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
         res.status(500).json(err)
     }
@@ -79,8 +91,12 @@ router.delete("/:id/:user", auth, async (req, res) => {
  */
 router.get("/find/:id", auth, async (req, res) => {
     try{
-        const course = await Course.findById(req.params.id)
-        res.status(200).json(course)
+        if(req.session.isAuth){
+            const course = await Course.findById(req.params.id)
+            res.status(200).json(course)
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
         res.status(500).json(err)
     }
@@ -93,9 +109,15 @@ router.get("/find/:id", auth, async (req, res) => {
  */
 router.get("/", auth, async (req, res) => {
     try{
-        const courses = await Course.find(req.query).populate('owner_id')
-        res.status(200).json(courses)
+        console.log(req.session)
+        if(req.session.isAuth){
+            const courses = await Course.find(req.query).populate('owner_id')
+            res.status(200).json(courses)
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
+        console.log(err)
         res.status(500).json(err)
     }
 })
@@ -107,14 +129,18 @@ router.get("/", auth, async (req, res) => {
  */
  router.get("/public", auth, async (req, res) => {
     try{
-        await Course.find({ $and:[{is_public: true},req.query]})
-        .exec(function(err, courses) {
-            if(err) {
-                console.log(err)
-            } else {
-                res.status(200).json(courses)
-            }
-        })
+        if(req.session.isAuth){
+            await Course.find({ $and:[{is_public: true},req.query]})
+            .exec(function(err, courses) {
+                if(err) {
+                    console.log(err)
+                } else {
+                    res.status(200).json(courses)
+                }
+            })
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     }catch(err){
         res.status(500).json(err)
     }
