+++ server/routes/Courses/courses_shared.js
@@ -11,8 +11,12 @@ const mongoose = require('mongoose');
 router.post("/", async (req, res) => {
     const newCourse = new CourseShared(req.body)
     try{
-        const savedCourse = await newCourse.save()
-        res.status(201).json(savedCourse)
+        if(req.session.isAuth){
+            const savedCourse = await newCourse.save()
+            res.status(201).json(savedCourse)
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     }catch(err){
         res.status(500).json(err)
     }
@@ -25,18 +29,22 @@ router.post("/", async (req, res) => {
  */
  router.get("/course/:course", async (req, res) => {
     try{
-        await CourseShared
-            .find({course_id: req.params.course})
-            .populate('roles')
-            .populate('course_id')
-            .populate('user_id')
-            .exec(function(err, courses) {
-                if(err) {
-                    console.log(err)
-                } else {
-                    res.status(200).json(courses)
-                }
-            })
+        if(req.session.isAuth){
+            await CourseShared
+                .find({course_id: req.params.course})
+                .populate('roles')
+                .populate('course_id')
+                .populate('user_id')
+                .exec(function(err, courses) {
+                    if(err) {
+                        console.log(err)
+                    } else {
+                        res.status(200).json(courses)
+                    }
+                })
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
         res.status(500).json(err)
     }
@@ -49,18 +57,22 @@ router.post("/", async (req, res) => {
  */
  router.get("/user/:user", async (req, res) => {
     try{
-        await CourseShared
-            .find({user_id: req.params.user})
-            .populate('roles')
-            .populate('course_id')
-            .populate('user_id')
-            .exec(function(err, courses) {
-                if(err) {
-                    console.log(err)
-                } else {
-                    res.status(200).json(courses)
-                }
-            })
+        if(req.session.isAuth){
+            await CourseShared
+                .find({user_id: req.params.user})
+                .populate('roles')
+                .populate('course_id')
+                .populate('user_id')
+                .exec(function(err, courses) {
+                    if(err) {
+                        console.log(err)
+                    } else {
+                        res.status(200).json(courses)
+                    }
+                })
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
         res.status(500).json(err)
     }
@@ -73,8 +85,12 @@ router.post("/", async (req, res) => {
  */
  router.delete("/:id", async (req, res) => {
     try{
-        await CourseShared.findByIdAndDelete(req.params.id)
-        res.status(200).json("The course shared has been deleted with this user")
+        if(req.session.isAuth){
+            await CourseShared.findByIdAndDelete(req.params.id)
+            res.status(200).json("The course shared has been deleted with this user")
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
         res.status(500).json(err)
     }
@@ -87,18 +103,22 @@ router.post("/", async (req, res) => {
  */
  router.get("/:user/:course", async (req, res) => {
     try{
-        await CourseShared
-        .find({user_id: req.params.user, course_id: req.params.course})
-        .populate('roles')
-        .populate('course_id')
-        .populate('user_id')
-        .exec(function(err, courses) {
-            if(err) {
-                console.log(err)
-            } else {
-                res.status(200).json(courses)
-            }
-        })
+        if(req.session.isAuth){
+            await CourseShared
+            .find({user_id: req.params.user, course_id: req.params.course})
+            .populate('roles')
+            .populate('course_id')
+            .populate('user_id')
+            .exec(function(err, courses) {
+                if(err) {
+                    console.log(err)
+                } else {
+                    res.status(200).json(courses)
+                }
+            })
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
         res.status(500).json(err)
     }
@@ -111,39 +131,42 @@ router.post("/", async (req, res) => {
  */
  router.post("/:user/:course", async (req, res) => {
     try{
-        var courseShared = await CourseShared.findOne({ course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user) })
+        if(req.session.isAuth){
+            var courseShared = await CourseShared.findOne({ course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user) })
 
-        var hasRole = await courseShared.roles.some(function (role) {
-            return role.equals(req.body.roles);
-        });
+            var hasRole = await courseShared.roles.some(function (role) {
+                return role.equals(req.body.roles);
+            });
 
-        // If the courseShared already has the role, we remove it
-        if(hasRole){
-            CourseShared.findOneAndUpdate(
-                { course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user) }, 
-                { $pull: { roles: req.body.roles } },
-               function (error, success) {
-                    if (error) {
-                        console.log(error);
-                    } else {
-                        res.redirect("/api/courses-shared/" + req.params.user + "/" + req.params.course)
+            // If the courseShared already has the role, we remove it
+            if(hasRole){
+                CourseShared.findOneAndUpdate(
+                    { course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user) }, 
+                    { $pull: { roles: req.body.roles } },
+                function (error, success) {
+                        if (error) {
+                            console.log(error);
+                        } else {
+                            res.redirect("/api/courses-shared/" + req.params.user + "/" + req.params.course)
+                        }
                     }
-                 }
-            );
-        } else { // if the courseShared doesn't have the role, we add it
-            CourseShared.findOneAndUpdate(
-                { course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user) }, 
-                { $addToSet: { roles: req.body.roles } },
-               function (error, success) {
-                    if (error) {
-                        console.log(error);
-                    } else {
-                        res.redirect("/api/courses-shared/" + req.params.user + "/" + req.params.course)
+                );
+            } else { // if the courseShared doesn't have the role, we add it
+                CourseShared.findOneAndUpdate(
+                    { course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user) }, 
+                    { $addToSet: { roles: req.body.roles } },
+                function (error, success) {
+                        if (error) {
+                            console.log(error);
+                        } else {
+                            res.redirect("/api/courses-shared/" + req.params.user + "/" + req.params.course)
+                        }
                     }
-                 }
-            );
+                );
+            }
+        } else {
+            res.status(500).json({"error": "connection-error"})
         }
-            
     } catch(err) {
         res.status(500).json(err)
     }
@@ -156,18 +179,22 @@ router.post("/", async (req, res) => {
  */
  router.post("/remove/:user/:course", async (req, res) => {
     try{
-        CourseShared.findOneAndUpdate(
-            {course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user)}, 
-            { $pull: { roles: req.body.roles  } },
-           function (error, success) {
-                if (error) {
-                    console.log(error);
-                } else {
-                    // res.json(success);
-                    res.redirect("/api/courses-shared/" + req.params.user + "/" + req.params.course)
+        if(req.session.isAuth){
+            CourseShared.findOneAndUpdate(
+                {course_id: mongoose.Types.ObjectId(req.params.course), user_id: mongoose.Types.ObjectId(req.params.user)}, 
+                { $pull: { roles: req.body.roles  } },
+                function (error, success) {
+                    if (error) {
+                        console.log(error);
+                    } else {
+                        // res.json(success);
+                        res.redirect("/api/courses-shared/" + req.params.user + "/" + req.params.course)
+                    }
                 }
-             }
-        );
+            )
+        } else {
+            res.status(500).json({"error": "connection-error"})
+        }
     } catch(err) {
         res.status(500).json(err)
     }
